Move utils to nestedCards.tsx.
